Move runner fixture to conftest.py
